+++ application/src/main/java/org/opentripplanner/routing/alternativelegs/AlternativeLegs.java
@@ -45,18 +45,10 @@ public class AlternativeLegs {
     Leg leg,
     Integer numberLegs,
     TransitService transitService,
-    boolean searchBackward,
+    NavigationDirection direction,
     AlternativeLegsFilter filter
   ) {
-    return getAlternativeLegs(
-      leg,
-      numberLegs,
-      transitService,
-      searchBackward,
-      filter,
-      false,
-      false
-    );
+    return getAlternativeLegs(leg, numberLegs, transitService, direction, filter, false, false);
   }
 
   /**
@@ -66,9 +58,8 @@ public class AlternativeLegs {
    * @param numberLegs           The number of alternative legs requested. If fewer legs are found,
    *                             only the found legs are returned.
    * @param transitService       The transit service used for the search
-   * @param includeDepartBefore  Boolean indicating whether the alternative legs should depart
-   *                             earlier or later than the original leg True if earlier, false if
-   *                             later.
+   * @param direction            Indicating whether the alternative legs should depart before or
+   *                             after than the original.
    * @param filter               AlternativeLegsFilter indicating which properties of the original
    *                             leg should not change in the alternative legs
    * @param exactOriginStop      Boolean indicating whether the exact departure stop of the original
@@ -82,7 +73,7 @@ public class AlternativeLegs {
     Leg leg,
     Integer numberLegs,
     TransitService transitService,
-    boolean includeDepartBefore,
+    NavigationDirection direction,
     AlternativeLegsFilter filter,
     boolean exactOriginStop,
     boolean exactDestinationStop
@@ -105,7 +96,7 @@ public class AlternativeLegs {
       ScheduledTransitLeg::getStartTime
     );
 
-    if (includeDepartBefore) {
+    if (direction == NavigationDirection.PREVIOUS) {
       legComparator = legComparator.reversed();
     }
 
@@ -119,13 +110,7 @@ public class AlternativeLegs {
       .distinct()
       .flatMap(tripPattern -> withBoardingAlightingPositions(origins, destinations, tripPattern))
       .flatMap(t ->
-        generateLegs(
-          transitService,
-          t,
-          leg.getStartTime(),
-          leg.getServiceDate(),
-          includeDepartBefore
-        )
+        generateLegs(transitService, t, leg.getStartTime(), leg.getServiceDate(), direction)
       )
       .filter(Predicate.not(leg::isPartiallySameTransitLeg))
       .sorted(legComparator)
@@ -142,7 +127,7 @@ public class AlternativeLegs {
     TripPatternBetweenStops tripPatternBetweenStops,
     ZonedDateTime departureTime,
     LocalDate originalDate,
-    boolean includeDepartBefore
+    NavigationDirection direction
   ) {
     TripPattern pattern = tripPatternBetweenStops.tripPattern;
     int boardingPosition = tripPatternBetweenStops.positions.boardingPosition;
@@ -155,7 +140,7 @@ public class AlternativeLegs {
       tts.getServiceDayMidnight() + tts.getRealtimeDeparture()
     );
 
-    if (includeDepartBefore) {
+    if (direction == NavigationDirection.PREVIOUS) {
       comparator = comparator.reversed();
     }
 
@@ -185,7 +170,7 @@ public class AlternativeLegs {
           continue;
         }
 
-        boolean departureTimeInRange = includeDepartBefore
+        boolean departureTimeInRange = direction == NavigationDirection.PREVIOUS
           ? tripTimes.getDepartureTime(boardingPosition) <= secondsSinceMidnight
           : tripTimes.getDepartureTime(boardingPosition) >= secondsSinceMidnight;